TFG: Prototipo finalizado completo

+++ backend/tareas-quarkus/src/main/java/es/upsa/dasi/tfg/tareas/application/usecases/impl/RemoveTareaByIdUsecaseImpl.java
@@ -2,7 +2,9 @@ package es.upsa.dasi.tfg.tareas.application.usecases.impl;
 
 import es.upsa.dasi.tfg.common.domain.exceptions.NotFoundTfgException;
 import es.upsa.dasi.tfg.tareas.application.TareaAuthorizationService;
+import es.upsa.dasi.tfg.tareas.application.TareaRulesService;
 import es.upsa.dasi.tfg.tareas.application.usecases.RemoveTareaByIdUsecase;
+import es.upsa.dasi.tfg.tareas.domain.model.Tarea;
 import es.upsa.dasi.tfg.tareas.domain.repository.Repository;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
@@ -12,19 +14,24 @@ public class RemoveTareaByIdUsecaseImpl implements RemoveTareaByIdUsecase
 {
     Repository repository;
     TareaAuthorizationService authz;
+    TareaRulesService rules;
 
     @Inject
-    public RemoveTareaByIdUsecaseImpl(Repository repository, TareaAuthorizationService authz)
+    public RemoveTareaByIdUsecaseImpl(Repository repository, TareaAuthorizationService authz, TareaRulesService rules)
     {
         this.repository = repository;
         this.authz = authz;
+        this.rules = rules;
     }
 
     @Override
     public void execute(long id) throws NotFoundTfgException
     {
-        long proyectoId = authz.proyectoIdOfTarea(id);
-        authz.requireMember(proyectoId);
+        Tarea tarea = repository.findById(id)
+                .orElseThrow(() -> new NotFoundTfgException("Tarea no encontrada: " + id));
+        long proyectoId = tarea.getProyectoId();
+        authz.requireProductOwner(proyectoId);
+        rules.validateCanDelete(tarea);
         repository.deleteById(id);
     }
 }
